+++ src/canary.rs
@@ -2,7 +2,11 @@ use std::time::Instant;
 
 use probe_rs::{Core, MemoryInterface, RegisterId};
 
-use crate::{registers::PC, Elf, TargetInfo, TIMEOUT};
+use crate::{
+    registers::PC,
+    target_info::{StackInfo, TargetInfo},
+    Elf, TIMEOUT,
+};
 
 /// Canary value
 const CANARY_U8: u8 = 0xAA;
@@ -11,40 +15,46 @@ const CANARY_U32: u32 = u32::from_le_bytes([CANARY_U8, CANARY_U8, CANARY_U8, CAN
 
 /// (Location of) the stack canary
 ///
-/// The stack canary is used to detect *potential* stack overflows
+/// The stack canary is used to detect *potential* stack overflows and report the
+/// amount of stack used.
 ///
-/// The canary is placed in memory as shown in the diagram below:
+/// The whole stack is initialized to `CANARY_U8` before the target program is started.
 ///
+/// When the programs ends (due to panic or breakpoint) the size of the canary is checked. If more
+/// than 90%  is "touched" (bytes != `CANARY_U8`) then that is considered to be a *potential* stack
+/// overflow. In any case the amount of used stack is reported.
+///
+/// Before execution:
 /// ``` text
-/// +--------+ -> initial_stack_pointer / stack_range.end()
+/// +--------+ -> stack_range.end() (initial_stack_pointer)
+/// |        |
+/// |        |
+/// | canary |
+/// |        |
+/// |        |
+/// +--------+ -> stack_range.start()
+/// | static | (variables, fixed size)
+/// +--------+ -> lowest RAM address
+/// ```
+///
+/// After execution:
+/// ``` text
+/// +--------+ -> stack_range.end() (initial_stack_pointer)
 /// |        |
 /// | stack  | (grows downwards)
 /// |        |
 /// +--------+
-/// |        |
-/// |        |
-/// +--------+
 /// | canary |
 /// +--------+ -> stack_range.start()
-/// |        |
 /// | static | (variables, fixed size)
-/// |        |
 /// +--------+ -> lowest RAM address
 /// ```
-///
-/// The whole canary is initialized to `CANARY_U8` before the target program is started.
-/// The canary size is 10% of the available stack space or 1 KiB, whichever is smallest.
-///
-/// When the programs ends (due to panic or breakpoint) the integrity of the canary is checked. If it was
-/// "touched" (any of its bytes != `CANARY_U8`) then that is considered to be a *potential* stack
-/// overflow.
 #[derive(Clone, Copy)]
 pub struct Canary {
-    address: u32,
-    size: usize,
-    stack_available: u32,
+    addr: u32,
     data_below_stack: bool,
-    measure_stack: bool,
+    size: u32,
+    size_kb: f64,
 }
 
 impl Canary {
@@ -53,142 +63,134 @@ impl Canary {
     /// Assumes that the target was reset-halted.
     pub fn install(
         core: &mut Core,
-        target_info: &TargetInfo,
         elf: &Elf,
-        measure_stack: bool,
-    ) -> Result<Option<Self>, anyhow::Error> {
-        let stack_info = match &target_info.stack_info {
+        target_info: &TargetInfo,
+    ) -> anyhow::Result<Option<Self>> {
+        let canary = match Self::prepare(elf, &target_info.stack_info) {
+            Some(canary) => canary,
+            None => return Ok(None),
+        };
+
+        let start = Instant::now();
+
+        // paint stack
+        paint_subroutine::execute(core, canary.addr, canary.size)?;
+
+        let seconds = start.elapsed().as_secs_f64();
+        canary.log_time("painting", seconds);
+
+        Ok(Some(canary))
+    }
+
+    /// Measure the stack usage.
+    ///
+    /// Returns `true` if a stack overflow is likely.
+    pub fn measure(self, core: &mut Core, elf: &Elf) -> anyhow::Result<bool> {
+        let start = Instant::now();
+
+        // measure stack usage
+        let touched_address = measure_subroutine::execute(core, self.addr, self.size)?;
+
+        let seconds = start.elapsed().as_secs_f64();
+        self.log_time("reading", seconds);
+
+        let min_stack_usage = match touched_address {
+            Some(touched_address) => {
+                log::debug!("stack was touched at {touched_address:#010X}");
+                elf.vector_table.initial_stack_pointer - touched_address
+            }
+            None => {
+                log::warn!("stack was not used at all");
+                0
+            }
+        };
+
+        let used_kb = min_stack_usage as f64 / 1024.0;
+        let pct = used_kb / self.size_kb * 100.0;
+        let msg = format!(
+            "program has used at least {used_kb:.2}/{:.2} KiB ({pct:.1}%) of stack space",
+            self.size_kb
+        );
+
+        // stack touched?
+        //
+        // We consider >90% stack usage a potential stack overflow
+        if pct > 90.0 {
+            log::warn!("{}", msg);
+            if self.data_below_stack {
+                log::warn!("data segments might be corrupted due to stack overflow");
+            }
+            Ok(true)
+        } else {
+            log::info!("{}", msg);
+            Ok(false)
+        }
+    }
+
+    /// Prepare, but not place the canary.
+    ///
+    /// If this succeeds, we have all the information we need in order to place the canary.
+    fn prepare(elf: &Elf, stack_info: &Option<StackInfo>) -> Option<Self> {
+        let stack_info = match stack_info {
             Some(stack_info) => stack_info,
             None => {
                 log::debug!("couldn't find valid stack range, not placing stack canary");
-                return Ok(None);
+                return None;
             }
         };
 
         if elf.program_uses_heap() {
             log::debug!("heap in use, not placing stack canary");
-            return Ok(None);
+            return None;
         }
 
-        let stack_start = *stack_info.range.start();
-        let stack_available = *stack_info.range.end() - stack_start;
-
-        let size = if measure_stack {
-            // When measuring stack consumption, we have to color the whole stack.
-            stack_available as usize
-        } else {
-            round_up(stack_available / 10, 4) as usize
-        };
+        let stack_addr = *stack_info.range.start();
+        let stack_size = *stack_info.range.end() - stack_addr;
 
         log::debug!(
-            "{stack_available} bytes of stack available ({:#010X} ..= {:#010X}), using {size} byte canary",
-            stack_info.range.start(),
+            "{stack_size} bytes of stack available ({stack_addr:#010X} ..= {:#010X})",
             stack_info.range.end(),
         );
 
-        let size_kb = size as f64 / 1024.0;
-        if measure_stack {
-            // Painting 100KB or more takes a few seconds, so provide user feedback.
-            log::info!("painting {size_kb:.2} KiB of RAM for stack usage estimation");
-        }
-        let start = Instant::now();
-        paint_subroutine::execute(core, stack_start, size as u32)?;
-        let seconds = start.elapsed().as_secs_f64();
-        log::trace!(
-            "setting up canary took {seconds:.3}s ({:.2} KiB/s)",
-            size_kb / seconds
-        );
+        Self::assert_subroutines(stack_addr, stack_size)?;
 
-        Ok(Some(Canary {
-            address: stack_start,
-            size,
-            stack_available,
+        Some(Canary {
+            addr: stack_addr,
             data_below_stack: stack_info.data_below_stack,
-            measure_stack,
-        }))
+            size: stack_size,
+            size_kb: stack_size as f64 / 1024.0,
+        })
     }
 
-    /// Detect if the stack canary was touched.
-    pub fn touched(self, core: &mut Core, elf: &Elf) -> anyhow::Result<bool> {
-        let size_kb = self.size as f64 / 1024.0;
-        if self.measure_stack {
-            log::info!("reading {size_kb:.2} KiB of RAM for stack usage estimation");
-        }
-        let start = Instant::now();
-        let touched_address = measure_subroutine::execute(core, self.address, self.size as u32)?;
-        let seconds = start.elapsed().as_secs_f64();
-        log::trace!(
-            "reading canary took {seconds:.3}s ({:.2} KiB/s)",
-            size_kb / seconds
-        );
-
-        let min_stack_usage = match touched_address {
-            Some(touched_address) => {
-                log::debug!("canary was touched at {touched_address:#010X}");
-                Some(elf.vector_table.initial_stack_pointer - touched_address)
-            }
-            None => None,
-        };
-
-        if self.measure_stack {
-            let min_stack_usage = min_stack_usage.unwrap_or(0);
-            let used_kb = min_stack_usage as f64 / 1024.0;
-            let avail_kb = self.stack_available as f64 / 1024.0;
-            let pct = used_kb / avail_kb * 100.0;
-            log::info!(
-                "program has used at least {used_kb:.2}/{avail_kb:.2} KiB ({pct:.1}%) of stack space"
-            );
-
-            // Don't test for stack overflows if we're measuring stack usage.
-            Ok(false)
-        } else {
-            match min_stack_usage {
-                Some(min_stack_usage) => {
-                    let used_kb = min_stack_usage as f64 / 1024.0;
-                    let avail_kb = self.stack_available as f64 / 1024.0;
-                    let pct = used_kb / avail_kb * 100.0;
-                    log::warn!(
-                        "program has used at least {used_kb:.2}/{avail_kb:.2} KiB ({pct:.1}%) of stack space",
-                    );
-
-                    if self.data_below_stack {
-                        log::warn!("data segments might be corrupted due to stack overflow");
-                    }
-
-                    Ok(true)
-                }
-                None => {
-                    log::debug!("stack canary intact");
-                    Ok(false)
-                }
-            }
-        }
+    fn log_time(&self, action: &str, seconds: f64) {
+        log::debug!(
+            "{action} {:.2} KiB of RAM took {seconds:.3}s ({:.2} KiB/s)",
+            self.size_kb,
+            self.size_kb / seconds
+        )
     }
-}
 
-/// Rounds up to the next multiple of `k` that is greater or equal to `n`.
-fn round_up(n: u32, k: u32) -> u32 {
-    match n % k {
-        0 => n,
-        rem => n + k - rem,
-    }
-}
-
-/// Assert 4-byte-alignment and that subroutine fits inside stack.
-macro_rules! assert_subroutine {
-    ($low_addr:expr, $stack_size:expr, $subroutine_size:expr) => {
-        assert_eq!($low_addr % 4, 0, "low_addr needs to be 4-byte-aligned");
-        assert_eq!($stack_size % 4, 0, "stack_size needs to be 4-byte-aligned");
+    /// Assert 4-byte-alignment and that subroutine fits inside stack.
+    fn assert_subroutines(stack_addr: u32, stack_size: u32) -> Option<()> {
+        assert_eq!(stack_addr % 4, 0, "low_addr needs to be 4-byte-aligned");
+        assert_eq!(stack_size % 4, 0, "stack_size needs to be 4-byte-aligned");
         assert_eq!(
-            $subroutine_size % 4,
+            paint_subroutine::size() % 4,
             0,
-            "subroutine needs to be 4-byte-aligned"
+            "paint subroutine needs to be 4-byte-aligned"
         );
-        assert!(
-            $subroutine_size < $stack_size,
-            "subroutine does not fit inside stack"
+        assert_eq!(
+            measure_subroutine::size() % 4,
+            0,
+            "measure subroutine needs to be 4-byte-aligned"
         );
-    };
+        if (stack_size < paint_subroutine::size()) || (stack_size < measure_subroutine::size()) {
+            log::warn!("subroutines do not fit in stack; not placing stack canary");
+            None
+        } else {
+            Some(())
+        }
+    }
 }
 
 /// Paint-stack subroutine.
@@ -238,7 +240,6 @@ mod paint_subroutine {
     /// paints the whole memory, except of itself. After the subroutine finishes
     /// executing we overwrite the subroutine using the probe.
     pub fn execute(core: &mut Core, low_addr: u32, stack_size: u32) -> Result<(), probe_rs::Error> {
-        assert_subroutine!(low_addr, stack_size, self::SUBROUTINE.len() as u32);
         super::execute_subroutine(core, low_addr, stack_size, self::SUBROUTINE)?;
         self::overwrite_subroutine(core, low_addr)?;
         Ok(())
@@ -259,6 +260,10 @@ mod paint_subroutine {
         0x00, 0xbe, // bkpt     0x0000
         0x00, 0xbe, // bkpt     0x0000 (padding instruction)
     ];
+
+    pub const fn size() -> u32 {
+        self::SUBROUTINE.len() as _
+    }
 }
 
 /// Measure-stack subroutine.
@@ -335,12 +340,9 @@ mod measure_subroutine {
         low_addr: u32,
         stack_size: u32,
     ) -> Result<Option<u32>, probe_rs::Error> {
-        assert_subroutine!(low_addr, stack_size, self::SUBROUTINE.len() as u32);
-
         // use probe to search through the memory the subroutine will be written to
-        match self::search_with_probe(core, low_addr)? {
-            addr @ Some(_) => return Ok(addr), // if we find a touched value, return early ...
-            None => {}                         // ... otherwise we continue
+        if let Some(addr) = self::search_with_probe(core, low_addr)? {
+            return Ok(Some(addr)); // return early, if we find a touched value
         }
 
         super::execute_subroutine(core, low_addr, stack_size, self::SUBROUTINE)?;
@@ -375,7 +377,7 @@ mod measure_subroutine {
             .to_le_bytes()
             .into_iter()
             .position(|b| b != CANARY_U8)
-            .unwrap();
+            .expect("some byte has to be touched, if `word_addr != 0`");
 
         Ok(Some(word_addr + offset as u32))
     }
@@ -392,13 +394,17 @@ mod measure_subroutine {
         0x00, 0xbe, // bkpt     0x0000
         0x00, 0xbe, // bkpt     0x0000 (padding instruction)
     ];
+
+    pub const fn size() -> u32 {
+        self::SUBROUTINE.len() as _
+    }
 }
 
 /// Prepare and run subroutine. Also clean up afterwards.
 ///
 /// # How?
 ///
-/// We place the parameters in the registers (see table below), place the subroutien
+/// We place the parameters in the registers (see table below), place the subroutine
 /// in memory, set the program counter to the beginning of the subroutine, execute
 /// the subroutine and reset the program counter afterwards.
 ///
@@ -439,23 +445,4 @@ fn execute_subroutine<const N: usize>(
     core.write_core_reg::<u32>(PC, previous_pc)?;
 
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use rstest::rstest;
-
-    #[rstest]
-    #[case(2, 4, 4)]
-    #[case(4, 4, 4)]
-    #[case(6, 4, 8)]
-    #[case(8, 4, 8)]
-    #[case::odd(5, 3, 6)]
-    #[should_panic]
-    #[case::div_zero(4, 0, 0)]
-    fn test_round_up(#[case] n: u32, #[case] k: u32, #[case] res: u32) {
-        assert_eq!(round_up(n, k), res);
-    }
 }