+++ src/main.rs
@@ -72,7 +72,7 @@ fn run_target_program(elf_path: &Path, chip_name: &str, opts: &cli::Opts) -> any
     // install stack canary
     let canary = match opts.no_reset {
         true => None, // cannot safely touch the stack of a running application
-        false => Canary::install(core, &target_info, elf, true)?,
+        false => Canary::install(core, elf, &target_info)?,
     };
     if canary.is_none() {
         log::info!("stack measurement was not set up");
@@ -80,19 +80,19 @@ fn run_target_program(elf_path: &Path, chip_name: &str, opts: &cli::Opts) -> any
 
     // run program and print logs until there is an exception
     attack_the_program(core, elf)?;
-    let current_dir = &env::current_dir()?;
+    let current_dir = env::current_dir()?;
     let halted_due_to_signal = print_logs(core, current_dir, elf, &target_info.memory_map, opts)?; // blocks until exception
     print_separator()?;
 
     // analyze stack canary
-    let canary_touched = canary
-        .map(|canary| canary.touched(core, elf))
+    let stack_overflow = canary
+        .map(|canary| canary.measure(core, elf))
         .transpose()?
         .unwrap_or(false);
 
     // print the backtrace
     let mut backtrace_settings =
-        backtrace::Settings::new(canary_touched, current_dir, halted_due_to_signal, opts);
+        backtrace::Settings::new(current_dir, halted_due_to_signal, opts, stack_overflow);
     let outcome = backtrace::print(core, elf, &target_info, &mut backtrace_settings)?;
 
     if !opts.no_reset {